Refactor code from Data namespace structure and file organization
